requirements.txt
Pillow==5.2.0
git+https://github.com/gjcarneiro/pybindgen.git#egg=pybindgen
